+++ src/org/usfirst/frc/team1554/lib/readonly/FloatConstant.java
@@ -42,7 +42,7 @@ public class FloatConstant implements ReadOnlyFloat {
 
     @Override
     public Float getValue() {
-        return Float.valueOf(this.value);
+        return this.value;
     }
 
     @Override
